[sw/ottf] Restructure meson build targets for OTTF.
For testing the OTTF, a single meson build target was constructed with
an example test built in. This refactors the build targets to build the
OTTF as a library that is linked with tests that are also defined as
libraries. This makes building test binaries that use the OTTF similar
to those that use the existing test_main.c test framework.

Additionally, two example OTTF tests (one bare-metal and one concurrency)
are added to provided examples of how to write on-device tests using the
OTTF.

This partially addresses a larger effort of refactoring the on-device
test framework, as described in #8015.
